+++ test/interfaces/file/mold_interface.cpp
@@ -137,7 +137,7 @@ SCENARIO("Dock file::Interface::filename", "[mold file::Interface]")
 
    system("rm -rf build/testarea");
    system("mkdir build/testarea");
-   system("cp test/etc/rebrand/librandom.sol build/testarea/librandom.so");
+   system("cp test/etc/retag/librandom.sol build/testarea/librandom.so");
 
    pathname = "build/testarea/";
    filename = "librandom.so";